refactor: FeedResult 폴더 구조 Command의 DTO 위치로 변경
