+++ src/main/java/com/example/bak/feed/application/command/FeedCommandService.java
@@ -1,8 +1,8 @@
 package com.example.bak.feed.application.command;
 
+import com.example.bak.feed.application.command.dto.FeedResult;
 import com.example.bak.feed.application.command.port.CommunityValidationPort;
 import com.example.bak.feed.application.command.port.FeedCommandPort;
-import com.example.bak.feed.application.query.dto.FeedResult;
 import com.example.bak.feed.domain.Feed;
 import com.example.bak.global.exception.BusinessException;
 import com.example.bak.global.exception.ErrorCode;
@@ -19,7 +19,6 @@ public class FeedCommandService {
     private final CommunityValidationPort communityValidationPort;
 
     public FeedResult createFeed(String title, String content, Long communityId, Long userId) {
-
         if (!communityValidationPort.isCommunityExists(communityId)) {
             throw new BusinessException(ErrorCode.COMMUNITY_NOT_FOUND);
         }
